restructioring of tab functions in to tabs.rs module
