+++ src/ui/tabs.rs
@@ -1,4 +1,10 @@
 use strum::{Display, EnumIter, FromRepr, IntoEnumIterator};
+use ratatui::{
+    layout::Rect,
+    style::{Color, Style, Stylize},
+    widgets::{Block, Borders, Tabs},
+    Frame,
+};
 
 #[derive(Default, Clone, Copy, Display, FromRepr, EnumIter)]
 pub enum SelectedTab {
@@ -26,4 +32,26 @@ impl SelectedTab {
         let next_index = current_index.saturating_add(1);
         Self::from_repr(next_index).unwrap_or(self)
     }
+}
+
+pub fn draw_tabs(frame: &mut Frame, area: Rect, selected_tab: &SelectedTab) {
+    let titles = SelectedTab::iter()
+        .map(|tab| {
+            let (name, color) = match tab {
+                SelectedTab::Tab1 => ("Typing Test", Color::Blue),
+                SelectedTab::Tab2 => ("Options", Color::Green),
+                SelectedTab::Tab3 => ("Account", Color::Magenta),
+                SelectedTab::Tab4 => ("About", Color::Red),
+            };
+            format!("  {}  ", name).fg(Color::White).bg(color)
+        })
+        .collect::<Vec<_>>();
+
+    let tabs = Tabs::new(titles)
+        .block(Block::default().borders(Borders::BOTTOM))
+        .highlight_style(Style::default().fg(Color::White).bg(Color::DarkGray))
+        .select((*selected_tab) as usize)
+        .divider("|");
+
+    frame.render_widget(tabs, area);
 }
